requirements.txt
asgiref==3.11.1
coverage==7.10.7
Django==4.2.28
djangorestframework==3.16.1
exceptiongroup==1.3.1
freezegun==1.5.5
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
psycopg==3.2.13
psycopg-binary==3.2.13
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
